Add files via upload
The first step is adding a new user in the database. If you don't have a username and password in the database, you will not access another page. If you start the application, it will show the login page. If you enter wrong username or password, it will show validation. If you let empty field, you will receive a message "This field is required." If you enter correct data, it will show you a new page for searching. Upon user submitting the query through a form submit button, the result of the query will get returned and displayed on the page.